requirements.txt
pyahocorasick
rich
pytsk3
